+++ client.go
@@ -443,7 +443,7 @@ func (client *Client) SendRawTransaction(rawData []byte) (hash types.Hash, err e
 	if e := tx.Decode(rawData, client.GetChainIDCached()); e != nil {
 		return "", errors.Wrap(e, "invalid raw transaction")
 	}
-	if tx.UnsignedTransaction.To.GetAddressType() == cfxaddress.AddressTypeUnknown {
+	if tx.UnsignedTransaction.To != nil && tx.UnsignedTransaction.To.GetAddressType() == cfxaddress.AddressTypeUnknown {
 		return "", errors.New("to address with unknown type is not allowed ")
 	}
 
@@ -690,6 +690,12 @@ func (client *Client) GetTransactionTraces(txHash types.Hash) (traces []types.Lo
 	return
 }
 
+// GetPosRewardByEpoch returns pos rewarded in this epoch
+func (client *Client) GetPosRewardByEpoch(epoch types.Epoch) (val *types.EpochPosReward, err error) {
+	err = client.CallRPC(&val, "cfx_getPoSRewardByEpoch", epoch)
+	return
+}
+
 // CreateUnsignedTransaction creates an unsigned transaction by parameters,
 // and the other fields will be set to values fetched from conflux node.
 func (client *Client) CreateUnsignedTransaction(from types.Address, to types.Address, amount *hexutil.Big, data []byte) (types.UnsignedTransaction, error) {
@@ -722,6 +728,7 @@ func (client *Client) ApplyUnsignedTransactionDefault(tx *types.UnsignedTransact
 
 	if client != nil {
 		if tx.From == nil {
+			//TODO: return error if client.AccountManager is nil?
 			if client.AccountManager != nil {
 				defaultAccount, err := client.AccountManager.GetDefault()
 				if err != nil {