+++ types/trace.go
@@ -12,13 +12,40 @@ import (
 )
 
 type TraceType string
+type CallType string
+type PocketType string
+type CreateType string
 
 const (
-	CALL_TYPE                      TraceType = "call"
-	CALL_RESULT_TYPE               TraceType = "call_result"
-	CREATE_TYPE                    TraceType = "create"
-	CREATE_RESULT_TYPE             TraceType = "create_result"
-	INTERNAL_TRANSFER_ACTIION_TYPE TraceType = "internal_transfer_action"
+	TRACE_CALL                      TraceType = "call"
+	TRACE_CALL_RESULT               TraceType = "call_result"
+	TRACE_CREATE                    TraceType = "create"
+	TRACE_CREATE_RESULT             TraceType = "create_result"
+	TRACE_INTERNAL_TRANSFER_ACTIION TraceType = "internal_transfer_action"
+)
+
+const (
+	CALL_NONE         CallType = "none"
+	CALL_CALL         CallType = "call"
+	CALL_CALLCODE     CallType = "callcode"
+	CALL_DELEGATECALL CallType = "delegatecall"
+	CALL_STATICCALL   CallType = "staticcall"
+)
+
+const (
+	POCKET_BALANCE                     PocketType = "balance"
+	POCKET_STAKING_BALANCE             PocketType = "staking_balance"
+	POCKET_STORAGE_COLLATERAL          PocketType = "storage_collateral"
+	POCKET_SPONSOR_BALANCE_FOR_GAS     PocketType = "sponsor_balance_for_gas"
+	POCKET_SPONSOR_BALANCE_FOR_STORAGE PocketType = "sponsor_balance_for_collateral"
+	POCKET_MINT_BURN                   PocketType = "mint_or_burn"
+	POCKET_GAS_PAYMENT                 PocketType = "gas_payment"
+)
+
+const (
+	CREATE_NONE    CreateType = "none"
+	CREATE_CREATE  CreateType = "create"
+	CREATE_CREATE2 CreateType = "create2"
 )
 
 type LocalizedBlockTrace struct {
@@ -36,6 +63,7 @@ type LocalizedTransactionTrace struct {
 
 type LocalizedTrace struct {
 	Action              interface{}     `json:"action"`
+	Valid               bool            `json:"valid"`
 	Type                TraceType       `json:"type"`
 	EpochHash           *Hash           `json:"epochHash,omitempty"`
 	EpochNumber         *hexutil.Big    `json:"epochNumber,omitempty"`
@@ -51,14 +79,15 @@ type Call struct {
 	Value    hexutil.Big   `json:"value"`
 	Gas      hexutil.Big   `json:"gas"`
 	Input    hexutil.Bytes `json:"input"`
-	CallType string        `json:"callType"`
+	CallType CallType      `json:"callType"`
 }
 
 type Create struct {
-	From  Address       `json:"from"`
-	Value hexutil.Big   `json:"value"`
-	Gas   hexutil.Big   `json:"gas"`
-	Init  hexutil.Bytes `json:"init"`
+	From       Address       `json:"from"`
+	Value      hexutil.Big   `json:"value"`
+	Gas        hexutil.Big   `json:"gas"`
+	Init       hexutil.Bytes `json:"init"`
+	CreateType CreateType    `json:"createType"`
 }
 
 type CallResult struct {
@@ -75,9 +104,11 @@ type CreateResult struct {
 }
 
 type InternalTransferAction struct {
-	From  Address     `json:"from"`
-	To    Address     `json:"to"`
-	Value hexutil.Big `json:"value"`
+	From       Address     `json:"from"`
+	FromPocket PocketType  `json:"fromPocket"`
+	To         Address     `json:"to"`
+	ToPocket   PocketType  `json:"toPocket"`
+	Value      hexutil.Big `json:"value"`
 }
 
 type TraceFilter struct {
@@ -107,44 +138,39 @@ func init() {
 // UnmarshalJSON unmarshals Input and Init type from []byte to hexutil.Bytes
 func (l *LocalizedTrace) UnmarshalJSON(data []byte) error {
 
+	type alias LocalizedTrace
+
+	a := alias{}
+	err := json.Unmarshal(data, &a)
+	if err != nil {
+		return err
+	}
+	*l = LocalizedTrace(a)
+
 	tmp := struct {
-		Action              map[string]interface{} `json:"action"`
-		Type                TraceType              `json:"type"`
-		EpochHash           *Hash                  `json:"epochHash"`
-		EpochNumber         *hexutil.Big           `json:"epochNumber"`
-		BlockHash           *Hash                  `json:"blockHash"`
-		TransactionPosition *hexutil.Uint64        `json:"transactionPosition"`
-		TransactionHash     *Hash                  `json:"transactionHash"`
+		Action map[string]interface{} `json:"action"`
 	}{}
 
-	err := json.Unmarshal(data, &tmp)
+	err = json.Unmarshal(data, &tmp)
 	if err != nil {
 		return err
 	}
 
-	l.Type = tmp.Type
-	l.EpochHash = tmp.EpochHash
-	l.EpochNumber = tmp.EpochNumber
-	l.BlockHash = tmp.BlockHash
-	l.TransactionPosition = tmp.TransactionPosition
-	l.TransactionHash = tmp.TransactionHash
-
 	var action interface{}
-	if action, err = parseAction(tmp.Action); err != nil {
+	if action, err = parseAction(tmp.Action, l.Type); err != nil {
 		return err
 	}
 	l.Action = action
-
 	return nil
 }
 
-func parseAction(actionInMap map[string]interface{}) (interface{}, error) {
-	actionKeys := utils.GetMapSortedKeys(actionInMap)
+func parseAction(actionInMap map[string]interface{}, actionType TraceType) (interface{}, error) {
+	// actionKeys := utils.GetMapSortedKeys(actionInMap)
 
-	newActionType := actionKeysToTypeMap[strings.Join(actionKeys, "")]
-	if newActionType == "" {
-		return nil, fmt.Errorf("uncongonized action type with fields %v", actionKeys)
-	}
+	// newActionType := actionKeysToTypeMap[strings.Join(actionKeys, "")]
+	// if newActionType == "" {
+	// 	return nil, fmt.Errorf("uncongonized action type with fields %v", actionKeys)
+	// }
 
 	actionJson, err := json.Marshal(actionInMap)
 	if err != nil {
@@ -152,38 +178,42 @@ func parseAction(actionInMap map[string]interface{}) (interface{}, error) {
 	}
 
 	var result interface{}
-	switch newActionType {
-	case "Call":
+	switch actionType {
+	case TRACE_CALL:
 		action := Call{}
 		err = json.Unmarshal(actionJson, &action)
 		result = action
-	case "Create":
+	case TRACE_CREATE:
 		action := Create{}
 		err = json.Unmarshal(actionJson, &action)
 		result = action
-	case "CallResult":
+	case TRACE_CALL_RESULT:
 		action := CallResult{}
 		err = json.Unmarshal(actionJson, &action)
 		result = action
-	case "CreateResult":
+	case TRACE_CREATE_RESULT:
 		action := CreateResult{}
 		err = json.Unmarshal(actionJson, &action)
 		result = action
-	case "InternalTransferAction":
+	case TRACE_INTERNAL_TRANSFER_ACTIION:
 		action := InternalTransferAction{}
 		err = json.Unmarshal(actionJson, &action)
 		result = action
+	default:
+		return nil, fmt.Errorf("unknown action type %v", actionType)
 	}
 	if err != nil {
-		return nil, errors.Wrapf(err, "failed to unmarshal %v to %v ", string(actionJson), newActionType)
+		return nil, errors.Wrapf(err, "failed to unmarshal %v to %v ", string(actionJson), actionType)
 	}
 
 	return result, nil
 }
 
-// TODO: should move to go sdk=========================
+type LocalizedTraceTire []*LocalizedTraceNode
+
 type LocalizedTraceNode struct {
 	Type                TraceType       `json:"type"`
+	Valid               bool            `json:"valid"`
 	EpochHash           *Hash           `json:"epochHash,omitempty"`
 	EpochNumber         *hexutil.Big    `json:"epochNumber,omitempty"`
 	BlockHash           *Hash           `json:"blockHash,omitempty"`
@@ -200,6 +230,7 @@ type LocalizedTraceNode struct {
 
 func (l *LocalizedTraceNode) populate(raw LocalizedTrace) {
 	l.Type = raw.Type
+	l.Valid = raw.Valid
 	l.EpochHash = raw.EpochHash
 	l.EpochNumber = raw.EpochNumber
 	l.BlockHash = raw.BlockHash
@@ -207,20 +238,35 @@ func (l *LocalizedTraceNode) populate(raw LocalizedTrace) {
 	l.TransactionHash = raw.TransactionHash
 }
 
-func TraceInTree(traces []LocalizedTrace) (node *LocalizedTraceNode, err error) {
+// TraceInTire convert flattened trace to trie, the convered tiers are stored in the order of the flattened trace like follow.
+//
+// InternalTransfer
+// call
+// create
+// createResult
+// callResult
+// InternalTransfer
+// ============>
+// InternalTransfer
+// call + callResult
+// 	|- call + callResult
+// InternalTransfer
+func TraceInTire(traces []LocalizedTrace) (tier LocalizedTraceTire, err error) {
 	cacheStack := new([]*LocalizedTraceNode)
 
 	for _, v := range traces {
-		if node == nil {
-			node, err = newLocalizedTraceNode(v, cacheStack)
+
+		if len(*cacheStack) == 0 {
+			n, err := newLocalizedTraceNode(v, cacheStack)
 			if err != nil {
 				return nil, errors.WithStack(err)
 			}
+			tier = append(tier, n)
 			continue
 		}
 
 		lastOpeningNode := (*cacheStack)[len(*cacheStack)-1]
-		if v.Type == INTERNAL_TRANSFER_ACTIION_TYPE {
+		if v.Type == TRACE_INTERNAL_TRANSFER_ACTIION {
 			item, err := newLocalizedTraceNode(v, nil)
 			if err != nil {
 				return nil, errors.WithStack(err)
@@ -229,7 +275,7 @@ func TraceInTree(traces []LocalizedTrace) (node *LocalizedTraceNode, err error) 
 			continue
 		}
 
-		if v.Type == CALL_TYPE || v.Type == CREATE_TYPE {
+		if v.Type == TRACE_CALL || v.Type == TRACE_CREATE {
 			item, err := newLocalizedTraceNode(v, cacheStack)
 			if err != nil {
 				return nil, errors.WithStack(err)
@@ -239,8 +285,8 @@ func TraceInTree(traces []LocalizedTrace) (node *LocalizedTraceNode, err error) 
 		}
 
 		// call result or create result
-		if lastOpeningNode.Type == CALL_TYPE {
-			if v.Type != CALL_RESULT_TYPE {
+		if lastOpeningNode.Type == TRACE_CALL {
+			if v.Type != TRACE_CALL_RESULT {
 				return nil, fmt.Errorf("expect trace type CallResult, got %v", v.Type)
 			}
 			cr := v.Action.(CallResult)
@@ -248,8 +294,8 @@ func TraceInTree(traces []LocalizedTrace) (node *LocalizedTraceNode, err error) 
 			*cacheStack = (*cacheStack)[:len(*cacheStack)-1]
 		}
 
-		if lastOpeningNode.Type == CREATE_TYPE {
-			if v.Type != CREATE_RESULT_TYPE {
+		if lastOpeningNode.Type == TRACE_CREATE {
+			if v.Type != TRACE_CREATE_RESULT {
 				return nil, fmt.Errorf("expect trace type CreateResult, got %v", v.Type)
 			}
 			cr := v.Action.(CreateResult)
@@ -259,13 +305,13 @@ func TraceInTree(traces []LocalizedTrace) (node *LocalizedTraceNode, err error) 
 	}
 	// push call trace, set to child when meet next call trace
 	// pop call trace when meet call result trace
-	return node, nil
+	return tier, nil
 }
 
 func newLocalizedTraceNode(trace LocalizedTrace, cacheStack *[]*LocalizedTraceNode,
 ) (*LocalizedTraceNode, error) {
 	switch trace.Type {
-	case CALL_TYPE:
+	case TRACE_CALL:
 		action := trace.Action.(Call)
 		node := &LocalizedTraceNode{CallWithResult: &TraceCallWithResult{
 			&action, nil,
@@ -273,7 +319,7 @@ func newLocalizedTraceNode(trace LocalizedTrace, cacheStack *[]*LocalizedTraceNo
 		node.populate(trace)
 		*cacheStack = append(*cacheStack, node)
 		return node, nil
-	case CREATE_TYPE:
+	case TRACE_CREATE:
 		action := trace.Action.(Create)
 		node := &LocalizedTraceNode{CreateWithResult: &TraceCreateWithResult{
 			&action, nil,
@@ -281,13 +327,20 @@ func newLocalizedTraceNode(trace LocalizedTrace, cacheStack *[]*LocalizedTraceNo
 		node.populate(trace)
 		*cacheStack = append(*cacheStack, node)
 		return node, nil
-	case INTERNAL_TRANSFER_ACTIION_TYPE:
+	case TRACE_INTERNAL_TRANSFER_ACTIION:
 		action := trace.Action.(InternalTransferAction)
 		node := &LocalizedTraceNode{InternalTransferAction: &action}
 		node.populate(trace)
 		return node, nil
 	}
 	return nil, fmt.Errorf("could not create new localized trace node by type %v", trace.Type)
+}
+
+func (t LocalizedTraceTire) Flatten() (flattened []*LocalizedTraceNode) {
+	for _, v := range t {
+		flattened = append(flattened, v.Flatten()...)
+	}
+	return flattened
 }
 
 func (l LocalizedTraceNode) Flatten() (flattened []*LocalizedTraceNode) {